create `zola index` command
create `zola index` command and route existing search indexing code to it

this is the first step in building a new `zola index` command that
allows building search indexes as a stand-alone task.

the new Subcommand in src/cli.rs outlines the planned api.

the biggest thing it will be providing compared to current code is
optional indexing with tantivy.

this commit doesn't add tantivy or any new functionality other than
allowing the existing search indexing to be performed as a subcommand.
(passing --index-type tantivy will trigger unimplemented!() panic)

All of this code was originally made by Jonathon Strong at
https://git.jstrong.dev/jstrong/zola/commit/cffb69c5bed085897d79f0ef0dda39ab32c8f840